pyproject.toml
[tool.poetry]
name = "manga-recommender-system"
version = "0.1.0"
description = ""
authors = ["Trung Dan Phan <trung-dan.phan@hec.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10,<3.12"
numpy = "1.24.3"
pandas = "2.0.3"
scikit-learn = "1.3.0"
scipy = "1.10.1"
matplotlib = "3.7.2"
seaborn = "0.12.2"
tqdm = "4.65.0"
spacy = "3.6.1"
gensim = "4.3.1"
sentence-transformers = "2.2.2"
optuna = "3.4.0"
mlflow = "2.6.0"
fastapi = "0.100.1"
flask = "2.3.3"
uvicorn = "0.23.2"
gunicorn = "20.1.0"
python-dotenv = "1.0.0"
requests = "2.31.0"
google-cloud-bigquery = "3.11.4"
db-dtypes = "^1.4.1"

[tool.poetry.group.dev.dependencies]
black = "23.7.0"
flake8 = "6.1.0"
isort = "5.12.0"
ipython = "8.14.0"
jupyterlab = "3.6.3"
notebook = "6.5.4"
loguru = "0.7.0"
mkdocs = "1.5.3"
typer = "0.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

